CLI 환경을 위한 cli_common, cli_async로 분리, 린트 해결
