+++ lib/common/cli_common.dart
@@ -1,0 +1,3 @@
+export 'package:fast_app_base/common/dart/extension/collection_extension.dart';
+export 'package:fast_app_base/common/dart/extension/num_duration_extension.dart';
+export 'package:fast_app_base/common/util/async/cli_async.dart';
