+++ lib/common/dart/extension/animation_controller_extension.dart
@@ -1,5 +1,5 @@
+import 'package:fast_app_base/common/cli_common.dart';
 import 'package:flutter/widgets.dart';
-import 'package:fast_app_base/common/common.dart';
 
 extension AnimationControllerExtension on AnimationController {
   void animateToTheEnd() {
