+++ lib/common/dart/extension/collection_extension.dart
@@ -1,0 +1,9 @@
+extension ListExtension<T> on List<T> {
+  void swap(int origin, int target) {
+    final temp1 = this[target];
+    this[target] = this[origin];
+    this[origin] = temp1;
+  }
+
+  Stream<T> toStream() => Stream.fromIterable(this);
+}
