+++ lib/common/util/async/cli_async.dart
@@ -1,18 +1,5 @@
 import 'dart:async';
 
-import 'package:flutter/material.dart';
-
-extension AsyncExt on Object {
-  Timer delay(Function func, [Duration duration = const Duration(milliseconds: 50)]) {
-    return Timer(duration, () {
-      if (this is State && !(this as State).mounted) {
-        return;
-      }
-      func();
-    });
-  }
-}
-
 Timer delayTimer(Function func, [Duration duration = const Duration(milliseconds: 50)]) {
   return Timer(duration, () {
     func();
@@ -35,12 +22,6 @@ Future sleepUntil(bool Function() predict,
   while (!predict()) {
     await sleepAsync(duration);
   }
-}
-
-void runOnUI(void Function() uiFunction) {
-  WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
-    uiFunction();
-  });
 }
 
 class DelayTarget {
