+++ lib/common/widget/round_button_theme.dart
@@ -11,12 +11,11 @@ enum RoundButtonTheme {
     this.bgColor,
     this.textColor,
     this.borderColor, {
-    this.shadowColor = Colors.transparent,
     this.backgroundColorProvider,
-  });
+  }) : shadowColor = Colors.transparent;
 
-  final Color Function(BuildContext context)?
-      backgroundColorProvider; //RoundButtonTheme 안에서 Custome Theme 분기가 필요하다면 이렇게 함수로 사용
+  ///RoundButtonTheme 안에서 Custome Theme 분기가 필요하다면 이렇게 함수로 사용
+  final Color Function(BuildContext context)? backgroundColorProvider;
   final Color bgColor;
   final Color textColor;
   final Color borderColor;
