+++ lib/screen/main/w_menu_drawer.dart
@@ -10,7 +10,6 @@ import '../../../screen/dialog/d_message.dart';
 import '../../common/common.dart';
 import '../../common/language/language.dart';
 import '../../common/theme/theme_util.dart';
-import '../../common/widget/w_empty_expanded.dart';
 import '../../common/widget/w_mode_switch.dart';
 
 class MenuDrawer extends StatefulWidget {
